Amplify to 0dB.
Because of the very limited signal-to-noise ratio, the audio input is now transformed to always make use of the full dynamic range. However, if the audio input is already "over-amplified", it is deliberately not quitened as it is often preferable to accept some distortion in order to get a higher signal level (and as such a better signal-to-noise ratio).